ReactJS Tasks - implementing simulation of ordering pizza. add: confirming action reconstruct: state of order froma array of dishes to object with order list as one of the fields add: isConfirmed field to state add: confirming order to basket container

+++ src/app/reducers/orderReducer.js
@@ -1,43 +1,61 @@
 import { historyTitle } from '../constants/constants'
 
-const order = (state = [], action) => {
+const order = (state = { orderList: [] }, action) => {
 
   switch (action.type) {
     case 'ADD_TO_ORDER': {
-      const matchedDish = state.find(dish => dish.id === action.dish.id)
+      const matchedDish = state.orderList.find(dish => dish.id === action.dish.id)
       
-      return matchedDish ? [
-        ...state.filter(dish => dish !== matchedDish),
-        { ...matchedDish, number: matchedDish.number + 1}
-      ]
-      : [
+      return {
         ...state,
-        { ...action.dish, number: 1}
-      ]
+        orderList: matchedDish ? [
+          ...state.orderList.filter(dish => dish !== matchedDish),
+          { ...matchedDish, number: matchedDish.number + 1}
+        ]
+        : [
+          ...state.orderList,
+          { ...action.dish, number: 1}
+        ]
+      }
     }
     case 'REMOVE_FROM_ORDER':{
-      const matchedDish = state.find(dish => dish.id === action.dish.id)
+      const matchedDish = state.orderList.find(dish => dish.id === action.dish.id)
 
-      return matchedDish.number > 1 ? [
-        ...state.filter(dish => dish !== action.dish),
-        { ...matchedDish, number: matchedDish.number - 1 }
-        
-      ].sort((dishA, dishB) => dishA.id - dishB.id) 
-      : [
-        ...state.filter(dish => dish !== action.dish)
-      ]
+      return {
+        ...state,
+        orderList: matchedDish.number > 1 ? [
+          ...state.orderList.filter(dish => dish !== action.dish),
+          { ...matchedDish, number: matchedDish.number - 1 }
+          
+        ].sort((dishA, dishB) => dishA.id - dishB.id) 
+        : [
+          ...state.orderList.filter(dish => dish !== action.dish)
+        ]
+      }
     }
     case 'MAKE_ORDER':
-      localStorage.getItem(historyTitle)
-        ? localStorage.setItem(historyTitle, JSON.stringify([
-          state,
-          ...JSON.parse(localStorage.getItem(historyTitle))
-        ]))
-        : localStorage.setItem(historyTitle, JSON.stringify([
-          state
-        ]))      
+      if (state.isConfirmed) {
+        localStorage.getItem(historyTitle)
+          ? localStorage.setItem(historyTitle, JSON.stringify([
+            state.orderList,
+            ...JSON.parse(localStorage.getItem(historyTitle))
+          ]))
+          : localStorage.setItem(historyTitle, JSON.stringify([
+            state.orderList
+          ]))      
 
-        return []
+          return {
+            ...state,
+            orderList: []
+          }
+      }
+        
+      return state
+    case 'GET_ORDER_CONFIRMATION':
+      return {
+        ...state,
+        isConfirmed: action.isConfirmed
+      }
     default:
       return state
   }
